#816 adding total and average row to status report table
